third_party/WebKit/LayoutTests/media/video-seek-to-current-position.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test that seeking attribute is true immediately after a seek
            to the current position, and goes back to false when the seek completes.</title>

        <script src="media-file.js"></script>
        <script src="video-test.js"></script>
        <script>
            var video;
            var seekedCount = 0;
            var seekTime = 1;

            function onload()
            {
                video = document.querySelector("video");
                video.src = findMediaFile("video", "content/test");
                waitForEventOnce('loadedmetadata', loadedmetadata);
                waitForEvent('seeking');
                waitForEvent('seeked', seeked);
            }

            function loadedmetadata()
            {
                run("video.currentTime = seekTime");
                testExpected("video.seeking", true);
                consoleWrite("");
            }

            function seeked()
            {
                ++seekedCount;
                consoleWrite("");

                testExpected("video.seeking", false);
                testExpected("video.currentTime", seekTime);

               if (seekedCount == 2) {
                   endTest();
                   return;
               }

               run("video.currentTime = seekTime");
               testExpected("video.seeking", true);

               consoleWrite("");
           }
        </script>
    </head>
    <body onload="onload()">
      <p>Test that seeking attribute is true immediately after a seek
        to the current position, and goes back to false when the seek completes.</p>
      <video></video>
    </body>
</html>
